validate findBy arguments to not be about association mappings.
