+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 
 * **2014-08-07**: properly handle fields mapped to nodename in DocumentRepository::findBy()
+  Report error when trying to find or order by association fields, instead of building
+  queries that won't work.
 
 1.2.0-RC1
 ---------
